+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Upcoming release
 
+### Breaking changes
+
+- semantics for `require` changed from `require('html-validate')` to
+  `require('html-validate').HtmlValidate` to support exposing other classes.
+
+### Features
+
+- new `TemplateExtractor` helper class for extracting templates from javascript
+  sources.
 - trigger downstream projects on release
 
 ## 0.8.3 (2018-06-12)